Fixed warnings + added Quaternion operators

+++ MathLibrary.cpp
@@ -281,23 +281,23 @@ void render()
 	glClear(GL_COLOR_BUFFER_BIT);
 
 	glColor3f(1.f, 0.f, 0.f);
-	for (int i = 0; i < casteljauPoints.size() - 1; i++)
+	for (unsigned int i = 0; i < casteljauPoints.size() - 1; i++)
 	{
 		glBegin(GL_LINES);
-		glVertex3f(casteljauPoints[i].getX(), casteljauPoints[i].getY(), casteljauPoints[i].getZ());
-		glVertex3f(casteljauPoints[i + 1].getX(), casteljauPoints[i + 1].getY(), casteljauPoints[i + 1].getZ());
+		glVertex3f((GLfloat)casteljauPoints[i].getX(), (GLfloat)casteljauPoints[i].getY(), (GLfloat)casteljauPoints[i].getZ());
+		glVertex3f((GLfloat)casteljauPoints[i + 1].getX(), (GLfloat)casteljauPoints[i + 1].getY(), (GLfloat)casteljauPoints[i + 1].getZ());
 		glEnd();
 	}
 
 	const int ANCHOR_SIZE = 2;
 	glColor3f(0.f, 0.f, 1.f);
-	for (int i = 0; i < bezierAnchors.size(); i++)
+	for (unsigned int i = 0; i < bezierAnchors.size(); i++)
 	{
 		glBegin(GL_QUADS);
-		glVertex3f(bezierAnchors[i].getX() - ANCHOR_SIZE, bezierAnchors[i].getY() - ANCHOR_SIZE, bezierAnchors[i].getZ());
-		glVertex3f(bezierAnchors[i].getX() + ANCHOR_SIZE, bezierAnchors[i].getY() - ANCHOR_SIZE, bezierAnchors[i].getZ());
-		glVertex3f(bezierAnchors[i].getX() + ANCHOR_SIZE, bezierAnchors[i].getY() + ANCHOR_SIZE, bezierAnchors[i].getZ());
-		glVertex3f(bezierAnchors[i].getX() - ANCHOR_SIZE, bezierAnchors[i].getY() + ANCHOR_SIZE, bezierAnchors[i].getZ());
+		glVertex3f((GLfloat)bezierAnchors[i].getX() - ANCHOR_SIZE, (GLfloat)bezierAnchors[i].getY() - ANCHOR_SIZE, (GLfloat)bezierAnchors[i].getZ());
+		glVertex3f((GLfloat)bezierAnchors[i].getX() + ANCHOR_SIZE, (GLfloat)bezierAnchors[i].getY() - ANCHOR_SIZE, (GLfloat)bezierAnchors[i].getZ());
+		glVertex3f((GLfloat)bezierAnchors[i].getX() + ANCHOR_SIZE, (GLfloat)bezierAnchors[i].getY() + ANCHOR_SIZE, (GLfloat)bezierAnchors[i].getZ());
+		glVertex3f((GLfloat)bezierAnchors[i].getX() - ANCHOR_SIZE, (GLfloat)bezierAnchors[i].getY() + ANCHOR_SIZE, (GLfloat)bezierAnchors[i].getZ());
 		glEnd();
 	}
